# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "School Scheduler",
    "version": "1.0"
  },
  "paths": {
...
    "/school/setup/work-week": {
      "post": {
        "tags": [
          "SchoolSetup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureWorkWeekCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureWorkWeekCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureWorkWeekCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "SchoolSetup"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/setup/lecture-slots": {
      "get": {
        "tags": [
          "SchoolSetup"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "SchoolSetup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLectureSlotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLectureSlotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLectureSlotCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/setup/break-slots": {
      "get": {
        "tags": [
          "SchoolSetup"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "SchoolSetup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBreakSlotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBreakSlotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBreakSlotCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/setup/lecture-slots/{id}": {
      "put": {
        "tags": [
          "SchoolSetup"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLectureSlotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLectureSlotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLectureSlotCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "SchoolSetup"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/setup/break-slots/{id}": {
      "put": {
        "tags": [
          "SchoolSetup"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBreakSlotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBreakSlotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBreakSlotCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "SchoolSetup"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/school/teachers/{teacherId}/availability": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{teacherId}/subjects": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/school/teachers": {
      "get": {
        "tags": [
          "TeacherManagement"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "TeacherManagement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}": {
      "put": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}/subjects": {
      "post": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSubjectsToTeacherCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSubjectsToTeacherCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSubjectsToTeacherCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}/subjects/{subjectId}": {
      "delete": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}/availability": {
      "post": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTeacherAvailabilityCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTeacherAvailabilityCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetTeacherAvailabilityCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}/availability/{dayOfWeek}/{slotNumber}": {
      "delete": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "dayOfWeek",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SchoolDayOfWeek"
            }
          },
          {
            "name": "slotNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/school/teachers/{id}/availability/clear": {
      "delete": {
        "tags": [
          "TeacherManagement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssignSubjectsToTeacherCommand": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BreakConfig": {
        "type": "object",
        "properties": {
          "afterSlotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
      "ConfigureWorkWeekCommand": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "dayConfigs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkWeekDayConfig"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateBreakSlotCommand": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/SchoolDayOfWeek"
          },
          "afterSlotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "CreateLectureSlotCommand": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/SchoolDayOfWeek"
          },
          "slotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "CreateTeacherCommand": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SchoolDayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SetTeacherAvailabilityCommand": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "availabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherAvailabilityConfig"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SlotConfig": {
        "type": "object",
        "properties": {
          "slotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
      "TeacherAvailabilityConfig": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "$ref": "#/components/schemas/SchoolDayOfWeek"
          },
          "slotNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBreakSlotCommand": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "afterSlotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateLectureSlotCommand": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "slotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTeacherCommand": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkWeekDayConfig": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "$ref": "#/components/schemas/SchoolDayOfWeek"
          },
          "lectureSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotConfig"
            },
            "nullable": true
          },
          "breakSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreakConfig"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}